width & height of the box/grid/thumbnail based on resolution table
now sizes are calculated automatically based on single config setting: thumbnail_size
which accepts integer: [1–9] - the larger, the smaller the thumbnail and the grid size.

Selected values are close as possible to the real resolution of the thumbnails.
To introduce as less visual image degradation as possible.
Result looking very close to the source image (as much as the size of the terminal's cell/row allows)

I think it is better that the thumbnail visually fits perfectly into the grid.
Without exceeding the grid size - subtle image blurriness is acceptable.
Perfect grid size was prioritized over pixel perfect thumbnail resolution.